Move Basic Auth class to sub-directory
